requirements.txt
Flask>=2.3.0
PyYAML>=6.0
pytest>=7.4.0
pytest-flask>=1.3.0
